requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
attrs==25.3.0
bcrypt==4.3.0
bleach==6.2.0
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
clamd==1.0.2
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
dash==3.0.4
dash-bootstrap-components==2.0.2
dash-uploader==0.6.1
dnspython==2.7.0
email_validator==2.2.0
fiona==1.10.1
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
Flask-WTF==1.2.2
GeoAlchemy2==0.17.1
geopandas==1.0.1
gevent==25.5.1
greenlet==3.2.2
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
minio==7.2.15
narwhals==1.40.0
nest-asyncio==1.6.0
numpy==2.0.2
packaging==25.0
pandas==2.2.3
passlib==1.7.4
plotly==6.1.0
psycopg2==2.9.10
pycparser==2.22
pycryptodome==3.23.0
pyogrio==0.11.0
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-magic==0.4.27
pytz==2025.2
requests==2.32.3
retrying==1.3.4
shapely==2.0.7
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
webencodings==0.5.1
Werkzeug==3.0.6
WTForms==3.2.1
zipp==3.21.0
zipstream-new==1.1.8
zope.event==5.0
zope.interface==7.2